requirements.txt
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.1

# Pinecone (new SDK) and text utilities
pinecone>=5.0.0
pinecone-text>=0.7.0

# Google AI SDKs (new and legacy for fallback)
google-genai>=0.1.0
google-generativeai>=0.8.3

# Enrichment utilities
beautifulsoup4>=4.12.0

# Optional fast JSON
orjson>=3.9.0